refactor: streamline app setup by removing unused functions and consolidating bindings
